refactor(actionbar): move cancel button into draggable container (partial backport)
